product admin views secured for admin users only

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -61,8 +62,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+@login_required
 def add_product(request):
     """A view to add a product to the store catalogue"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admins can perform this action')
+        return redirect(redirect('home'))
+        
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -81,8 +87,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """A view to edit a product in the store catalogue"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admins can perform this action')
+        return redirect(redirect('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     
     if request.method == 'POST':
@@ -106,9 +117,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store catalogue"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only admins can perform this action')
+        return redirect(redirect('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
